+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Development
 
 - Add test coverage report
 - Make consistent use of original_table_id/test_table_id
+- Add tests for is_sql
 
 0.2.0 (2023-06-12)
 ******************